appproject/requirements_temp.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
awscli==1.41.0
boto3==1.39.0
botocore==1.39.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.2.1
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
cycler==0.12.1
datetimejp==1.0.0
debugpy==1.8.14
decorator==5.2.1
dill==0.4.0
Django==5.2.6
docker==7.1.0
docutils==0.19
executing==2.2.0
fastapi==0.115.13
fonttools==4.53.1
google-pasta==0.2.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
h11==0.16.0
idna==3.8
importlib-metadata==6.11.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jmespath==1.0.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mock==4.0.3
multiprocess==0.70.18
nest-asyncio==1.6.0
numpy==1.26.4
omegaconf==2.3.0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pathos==0.3.4
pillow==10.4.0
platformdirs==4.3.8
pox==0.3.6
ppft==1.7.7
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
psycopg2==2.9.10
pure_eval==0.2.3
pyasn1==0.6.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2024.2
pywin32==310
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rpds-py==0.25.1
rsa==4.7.2
s3transfer==0.13.0
sagemaker==2.247.1
sagemaker-core==1.0.40
schema==0.7.7
scipy==1.16.0
six==1.16.0
smdebug-rulesconfig==1.0.1
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
tblib==3.1.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2024.2
urllib3==2.2.2
uvicorn==0.34.3
wcwidth==0.2.13
xgboost==3.0.2
zipp==3.23.0
channels==4.1.0
daphne==4.0.0
httpx==0.24.1